requirements.txt
aiogram>=3.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.2
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
lxml>=5.0.0
openai>=1.0.0
